Revert "Introduce ASTs for arguments of macro expansions (#1438)"
This reverts commit a49d86f8f4b42ce0b1a336f50bc361c71e038604.
