+++ c2cpg-tests/src/test/scala/io/shiftleft/c2cpg/querying/CDataFlowTests.scala
@@ -1,7 +1,8 @@
 package io.shiftleft.c2cpg.querying
 
 import io.shiftleft.c2cpg.testfixtures.DataFlowCodeToCpgSuite
-import io.shiftleft.codepropertygraph.generated.EdgeTypes
+import io.shiftleft.codepropertygraph.generated.nodes.Call
+import io.shiftleft.codepropertygraph.generated.{DispatchTypes, EdgeTypes}
 import io.shiftleft.dataflowengineoss.language._
 import io.shiftleft.semanticcpg.language._
 import overflowdb.traversal.Traversal
@@ -977,5 +978,43 @@ class CDataFlowTests32 extends DataFlowCodeToCpgSuite {
     sink.size shouldBe 2
     source.size shouldBe 2
     sink.reachableBy(source).size shouldBe 4
+  }
+}
+
+class CfgMacroTests extends DataFlowCodeToCpgSuite {
+  override val code: String =
+    """
+       #define MP4_GET4BYTES( dst ) MP4_GETX_PRIVATE( dst, GetDWBE(p_peek), 4 )
+       #define MP4_GETX_PRIVATE(dst, code, size) \
+    do \
+    { \
+        if( (i_read) >= (size) ) \
+        { \
+            dst = (code); \
+            p_peek += (size); \
+            i_read -= (size); \
+        } \
+        else \
+        { \
+            dst = 0; \
+            i_read = 0; \
+        } \
+    } while(0)
+
+    int foo() {
+       unsigned int x;
+       MP4_GET4BYTES(x);
+       sink(x);
+    }
+
+    """.stripMargin
+
+  "should create correct CFG for macro expansion and find data flow" in {
+    val List(callToMacro: Call) = cpg.method("foo").call.dispatchType(DispatchTypes.INLINED).l
+    callToMacro.argument.code.l shouldBe List("x")
+    callToMacro.cfgNext.code.toSet shouldBe Set("x", "i_read")
+    val source = cpg.method("foo").call.name("MP4_GET4BYTES").argument(1).l
+    val sink = cpg.method("foo").call.name("sink").argument(1).l
+    sink.reachableByFlows(source).l.foreach(println)
   }
 }
