+++ c2cpg/src/main/scala/io/shiftleft/c2cpg/astcreation/MacroHandler.scala
@@ -1,7 +1,7 @@
 package io.shiftleft.c2cpg.astcreation
 
 import io.shiftleft.codepropertygraph.generated.DispatchTypes
-import io.shiftleft.codepropertygraph.generated.nodes.{AstNodeNew, NewCall, NewLocal}
+import io.shiftleft.codepropertygraph.generated.nodes.{NewCall, NewIdentifier}
 import io.shiftleft.x2cpg.Ast
 import org.eclipse.cdt.core.dom.ast.{IASTMacroExpansionLocation, IASTNode, IASTPreprocessorMacroDefinition}
 import org.eclipse.cdt.internal.core.parser.scanner.MacroArgumentExtractor
@@ -16,16 +16,16 @@ trait MacroHandler {
   /**
     * For the given node, determine if it is expanded from a macro, and if so,
     * create a Call node to represent the macro invocation and attach `ast`
-    * as its child. Also create a METHOD node to represent the macro if it has
-    * not been created as part of processing the current translation unit.
-    * Removal of duplicates that result from usage of the same macro in multiple
-    * compilation units is deferred to a pass that runs once all ASTs have been
-    * created.
+    * as its child.
     * */
   def asChildOfMacroCall(node: IASTNode, ast: Ast, order: Int): Ast = {
     val macroCallAst = extractMatchingMacro(node).map {
       case (mac, args) =>
-        createMacroCallAst(ast, node, mac, args, order)
+        createMacroCallAst( // ast,
+                           node,
+                           mac,
+                           args,
+                           order)
     }
     if (macroCallAst.isDefined) {
       // TODO order/argument_index of `ast`'s root needs to be set to 1
@@ -43,12 +43,14 @@ trait MacroHandler {
     * */
   private def extractMatchingMacro(node: IASTNode): Option[(IASTPreprocessorMacroDefinition, List[String])] = {
     expandedFromMacro(node)
-      .filterNot { m =>
-        isExpandedFrom(node.getParent, m)
+      .filterNot { expandedFrom =>
+        expandedFromMacro(node.getParent)
+          .map(_.getExpansion.getMacroDefinition)
+          .contains(expandedFrom.getExpansion.getMacroDefinition)
       }
-      .foreach { m =>
+      .foreach { expandedFrom =>
         val nodeOffset = node.getFileLocation.getNodeOffset
-        val macroName = m.getExpansion.getMacroDefinition.getName.toString
+        val macroName = expandedFrom.getExpansion.getMacroDefinition.getName.toString
         while (nodeOffsetMacroPairs.headOption.exists(x => x._1 <= nodeOffset)) {
           val (_, macroDefinition) = nodeOffsetMacroPairs.head
           nodeOffsetMacroPairs.remove(0)
@@ -63,32 +65,12 @@ trait MacroHandler {
   }
 
   /**
-    * Determine whether `node` is expanded from the macro expansion
-    * at `loc`.
-    * */
-  private def isExpandedFrom(node: IASTNode, loc: IASTMacroExpansionLocation) = {
-    expandedFromMacro(node)
-      .map(_.getExpansion.getMacroDefinition)
-      .contains(loc.getExpansion.getMacroDefinition)
-  }
-
-  private def argumentTrees(arguments: List[String], ast: Ast): List[Option[Ast]] = {
-    arguments.map { arg =>
-      val rootNode =
-        ast.nodes.find(x => !x.isInstanceOf[NewLocal] && x.properties.get("CODE").contains(arg.replace(" ", "")))
-      rootNode.map { x =>
-        ast.subTreeCopy(x.asInstanceOf[AstNodeNew], 1)
-      }
-    }
-  }
-
-  /**
     * Create an AST that represents a macro expansion as a call.
     * The AST is rooted in a CALL node and contains sub trees
     * for arguments. These are also connected to the AST via
     * ARGUMENT edges.
     * */
-  private def createMacroCallAst(ast: Ast,
+  private def createMacroCallAst( // ast: Ast,
                                  node: IASTNode,
                                  macroDef: IASTPreprocessorMacroDefinition,
                                  arguments: List[String],
@@ -106,10 +88,22 @@ trait MacroHandler {
       .order(order)
       .argumentIndex(order)
 
-    val argAsts = argumentTrees(arguments, ast).map(_.getOrElse(Ast()))
+    // TODO We want to clone the ASTS of arguments here
+    // and then attach those ASTS to the AST we return
+    // For now, we instead create identifier nodes
+    // val nodes = ast.nodes
+    // val argRoots = arguments.flatMap { arg =>
+    //  nodes.find(x => x.properties.get("CODE").contains(arg))
+    // }
+
+    val argAsts = arguments.zipWithIndex.map {
+      case (arg, i) =>
+        Ast(NewIdentifier().name(arg).code(arg).order(i + 1).argumentIndex(i + 1))
+    }
+
     Ast(callNode)
       .withChildren(argAsts)
-      .withArgEdges(callNode, argAsts.flatMap(x => x.root))
+      .withArgEdges(callNode, argAsts.flatMap(_.root))
   }
 
   private val nodeOffsetMacroPairs: mutable.Buffer[(Int, IASTPreprocessorMacroDefinition)] = {
